Add sort by Project Name (in issue url) #71
Fixes #71
Add projectname in options in root.js to filter the claims.
Add filterproj in root.js to get the claims according to their
project.
Modify whereClause in datautils.js to check for projectname or
username
if they are available.
Add distinctProjects in datautils.js to get distinct projects from
database.
Change Placeholder in view.hbs to add search with username and
projectname.
Add Scrolling Menu for project names and modify the page navigation
at the bottom for both username and projectname.

Signed-off-by: Anshul Mittal <anshulmittal0409@gmail.com>